\name{dmr.linreg-package}
\alias{dmr.linreg-package}
\alias{dmr.linreg}

\docType{package}

\title{A package with example functions from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{A package with example functions from Chapter 8 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\details{See examples in Chapter 8.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{models}

\seealso{
\code{\link[dmr.util:dmr.util-package]{dmr.util}}
\code{\link[dmr.regeval:dmr.regeval-package]{dmr.regeval}}
\code{\link[dmr.linclas:dmr.linclas-package]{dmr.linclas}}
}

\examples{
}
