\name{lrdat}
\alias{lrdat}
\alias{lrdat.train}
\alias{lrdat.test}
\alias{prdat}
\alias{prdat.train}
\alias{prdat.test}

\docType{data}

\title{Example datasets from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{Example datasets from Chapter 8 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\usage{See Section 8.1, Example 8.1.1.}

\details{See Section 8.1, Example 8.1.1.}

\source{
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}
\examples{
data(lrdat)
data(lrdat.train)
data(lrdat.test)
summary(prdat)
summary(prdat.train)
summary(prdat.test)
}

\keyword{datasets}
